pyproject.toml
[tool.poetry]
name = "airshot-plot"
version = "0.1.0"
description = ""
authors = ["TsukasaHori <best.tsukasa@gmail.com>"]
readme = "README.md"
packages = [{include = "airshot_plot"}]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.24.3"
staticmap = "^0.5.5"
matplotlib = "^3.7.1"
pillow = "^9.5.0"
pandas = "^2.0.2"
chardet = "^5.1.0"
pyproj = "^3.6.0"
configparser = "^6.0.0"
openpyxl = "^3.1.2"
quaternion = {git = "https://github.com/moble/quaternion.git"}
numpy-quaternion = "^2022.4.3"
numba = "^0.57.1"
findiff = "^0.10.0"
folium = "^0.15.0"
ray = "^2.9.0"
simplejson = "^3.19.2"
control = "^0.9.4"
simplekml = "^1.3.6"


[tool.poetry.group.dev.dependencies]
mypy = "^1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
